changed machine to 6 as 1-5 were issued to other people
