+++ src/interactions/map_arcade/game_machine_2.interactions.js
@@ -1,4 +1,3 @@
-import { scaleFactor } from '../../constants';
 import { displayDialogue } from '../../utils';
 
 export const interactionWithGameMachine2 = (player, k, map) => {
@@ -86,114 +85,5 @@ function closeCustomPrompt() {
 // Function to start the Chrome Dino Game
 function startChromeDinoGame(k) {
     k.debug.log('Chrome Dino Game started!');
-
-    //variables for different things in the game
-    const FLOOR_HEIGHT = 48;
-    const JUMP_FORCE = 800;
-    const SPEED = 480;
-    const GRAVITY = 1600
-
-    // Set up the game scene
-    k.scene('dinoGame', () => {
-        //gravity for dino
-        k.setGravity(GRAVITY);
-
-        //load a sprite name dino with run animation
-        k.loadSprite('dino', 'public/doux.png', {
-            sliceX: 24,
-            sliceY: 1,
-            anims: {
-                run: { from: 17, to: 23, loop: true, speed: 10 },
-            },
-        });
-
-        //add the dino to screen
-        const dino = k.add([
-            k.sprite('dino', { anim: 'run' }),
-            k.pos(80, 200),
-            k.area(),
-            k.body(),
-            k.scale(scaleFactor),
-        ]);
-
-        //pressing the spacebar lets the dino jump
-        k.onKeyPress('space', () => {
-            if (dino.isGrounded()) {
-                dino.jump(JUMP_FORCE);
-            }
-        });
-
-        //add platform
-        k.add([
-            k.rect(k.width(), FLOOR_HEIGHT),
-            k.pos(0, k.height() - 250),
-            k.anchor('botleft'),
-            k.outline(4),
-            k.area(),
-            k.body({ isStatic: true }),
-            k.color(127, 200, 255),
-        ]);
-
-        //spawn trees with different heights at different intervals
-        function spawnTree() {
-            k.add([
-                k.rect(48, k.rand(48, 72)),
-                k.area(),
-                k.outline(4),
-                k.pos(k.width() - 10, k.height() - 500),
-                k.anchor('botleft'),
-                k.body(),
-                k.color(255, 180, 255),
-                k.move(k.LEFT, SPEED),
-                'tree',
-            ]);
-
-            k.wait(k.rand(1.5, 3), () => {
-                spawnTree();
-            });
-        }
-        spawnTree();
-
-        //on hit show an animation and change scenes
-        dino.onCollide('tree', () => {
-            k.addKaboom(dino.pos);
-            k.shake();
-            k.go('lose', score); // go to "lose" scene here
-        });
-
-        // keep track of score
-        let score = 0;
-        const scoreLabel = k.add([k.text(score), k.pos(68, 100)]);
-        // increment score every frame
-        k.onUpdate(() => {
-            score++;
-            scoreLabel.text = score;
-        });
-    });
-
-    //set up lose screen
-    k.scene('lose', (score) => {
-        k.setGravity(0);
-        //add "game over" text
-        k.add([k.text('Game Over'), k.pos(k.center()), k.anchor('center')]);
-        // display score
-        k.add([
-            k.text(score),
-            k.pos(k.width() / 2, k.height() / 2 + 80),
-            k.scale(2),
-            k.anchor('center'),
-        ]);
-
-        // go back to game with space is pressed
-        k.onKeyPress('space', () => import('../../scenes/arcade').then((_) => {
-            k.go('arcade');
-        }));
-        k.onClick(() => import('../../scenes/arcade').then((_) => {
-            k.go('arcade');
-        }));
-        k.setGravity(0);
-    });
-
-    // Start the game scene
-    k.go('dinoGame');
+    // Here you can implement the logic to actually start the game
 }
