+++ src/interactions/map_arcade/index.js
@@ -1,5 +1,6 @@
 import { enterMapCityInteraction } from './enterMapCity.interactions';
 import { interactionWithGameMachine2 } from './game_machine_2.interactions';
+import { interactionWithGameMachine6 } from './game_machine_6.interactions';
 
 const interactions = [
     enterMapCityInteraction,
@@ -7,6 +8,7 @@ const interactions = [
 
     // new interaction
     interactionWithGameMachine2,
+    interactionWithGameMachine6,
 ];
 
 export const attachInteractions = (gameObj, k) => {
